Removed warning caused by calling std::min<int> with size_t values

+++ Source/Core/Core/DSP/DSPCodeUtil.cpp
@@ -60,7 +60,7 @@ bool Compare(const std::vector<u16> &code1, const std::vector<u16> &code2)
 	if (code1.size() != code2.size())
 		printf("Size difference! 1=%i 2=%i\n", (int)code1.size(), (int)code2.size());
 	u32 count_equal = 0;
-	const int min_size = std::min<int>(code1.size(), code2.size());
+	const int min_size = std::min<int>((int)code1.size(), (int)code2.size());
 
 	AssemblerSettings settings;
 	DSPDisassembler disassembler(settings);
